==> notebooks/01-data-preparation/0x-extract-englewood-boundaries-ward-and-community-areas.ipynb <==
# Extracting Greater Englewood from Community Area Shapefile
so I can identify overlap of wards with community areas
2/25/24

import json

### get Englewood community areas

# # Load the GeoJSON file
# with open('../data/01-raw/Boundaries - Community Areas (current).geojson', 'r') as file:
#     geojson_ccas = json.load(file)

# # List of community area names or IDs to keep
# areas_to_keep = ['ENGLEWOOD', 'WEST ENGLEWOOD']

# # Filter to include only areas_to_keep
# selected_features = [feature for feature in geojson_ccas['features'] if feature['properties']['community'] in areas_to_keep]

# # Update the GeoJSON data to include only those selected
# geojson_ccas['features'] = selected_features

# # Write the modified GeoJSON data back to a new file
# with open('../data/01-raw/englewood_communities.geojson', 'w') as file:
#     json.dump(geojson_ccas, file)

### get Englewood wards

# # Load the GeoJSON file
# with open('../data/01-raw/Boundaries - Wards (2023-)_20240225.geojson', 'r') as file:
#     geojson_wards = json.load(file)

# # List of community area names or IDs to keep
# areas_to_keep = ['6', '15', '16', '17', '20']

# # Filter to include only areas_to_keep
# selected_features = [feature for feature in geojson_wards['features'] if feature['properties']['ward_id'] in areas_to_keep]

# # Update the GeoJSON data to include only those selected
# geojson_wards['features'] = selected_features

# # Write the modified GeoJSON data back to a new file
# with open('../data/01-raw/englewood_wards.geojson', 'w') as file:
#     json.dump(geojson_wards, file)

### get Englewood old wards

# Load the GeoJSON file
with open('../../gis/shapefiles/Boundaries - Wards (2015-2023).geojson', 'r') as file:
    geojson_wards = json.load(file)

# List of community area names or IDs to keep
areas_to_keep = ['6', '15', '16', '17', '20']

# Filter to include only areas_to_keep
selected_features = [feature for feature in geojson_wards['features'] if feature['properties']['ward'] in areas_to_keep]

# Update the GeoJSON data to include only those selected
geojson_wards['features'] = selected_features

# Write the modified GeoJSON data back to a new file
with open('../../gis/shapefiles/englewood_wards_old.geojson', 'w') as file:
    json.dump(geojson_wards, file)
