==> notebooks/01-data-preparation/01b-tidy-census-by-community.ipynb <==
import pandas as pd

df = pd.read_csv("../../data/00-raw/population_cmap_2022.csv")

df

# df.dtypes
for column in df.columns:
    print(f"{column}: {df[column].dtype}")

### keep only worthwhile fields

df_census = df[['GEOG','2020_POP','VAC_HU','HU_TOT','VACANT','VACperc']]
df_census['community_caps']=df_census['GEOG'].str.upper()
df_census.columns = df_census.columns.str.lower()
df_census = df_census.rename(columns={'geog':'community_name'})
df_census

### Replace 'The Loop' with 'Loop', to match community names in other data sets

# boundaries shapefile refers to downtown as "LOOP", not "The Loop"
df_census['community_name']= df_census['community_name'].replace('The Loop','Loop')
df_census['community_caps']= df_census['community_caps'].replace('THE LOOP','LOOP')
df_census

df_census[df_census['community_caps']=='THE LOOP']

# Export

df_census.to_csv("../../data/01-tidied/census-by-community.csv", index=False)
