==> notebooks/01-data-preparation/01c-tidy-unshoveled-sidewalk-fines.ipynb <==
# Tidy Unshoveled Sidewalk Fines
3/2/24

This notebook organizes and tidies FOIA-requested datasets provided by the Department of Administrative Hearings:
1) One dataset of all court dockets for unshoveled sidewalks, including data on issuing department
2) One dataset of the same court dockets, including additional data on hearing date, case disposition, respondents, etc.

My goal is to join this into one comprehensive (albeit denormalized) dataset, dockets-to-details.csv
and three normalized datasets:

Though a full year of data is provided for the 2018-2019 season, I chose to look only at July 2019- June 2023 to align with data available for 311 complaints as well as City Council's legislative session running from May 2019 to May 2023.<br>

### summary of recordsets

Following is a summary of all recordsets prepared in this notebook, along with a count after filtering for 7/1/19 to 6/30/23

|recordset|num records|cardinality |description|
|---|---|---|---|
| dockets-to-addresses.csv | 1918 |  1:1 |docket numbers and cleaned addresses<br>to facilitate geocoding |
| dockets-to-depts.csv | 1917 |  1:1 |docket numbers to departments,<br>one record with missing dept |
| dockets-to-respondents.csv | 2240 | 1:many | one record per docket per respondent |
| dockets-detailed.csv | 3662 | 1:many | full details for each docket |

### notes on data quality
* one docket had no violation date (21DT00019A). Rather than estimate this date, I filtered it out
* one docket had no info about which department issued the citation. I retained this in dockets-to-details

The following dockets could not be geocoded. As a result, all summaries by ward and by community area will have 1913 records, rather than 1918
* four dockets have address listed as UNKNOWN. This was not discovered until the geocoding phase
* one docket contains a nonsensical address- 5450 S 47TH

### overview of this data preparation
<ol>
<li>Read Data
<li>Prep Dockets Data (Simple)
<li>Prep Dockets Data (Detailed)
<li>Normalize Data
<li>Prep Addresses to Facilitate Geocoding
<li>Merge Department Info into Detailed Dockets Dataset
<li>Export Data
</ol>

# Read Data

import pandas as pd

df_dockets = pd.read_csv("../../data/00-raw/FOIA-H064920-011124-snow-dockets.csv")
df_dockets_detailed = pd.read_csv("../../data/00-raw/FOIA-H067483-021424-snow-dockets-detailed.csv")
df_corrections = pd.read_csv("../../data/00-raw/address-corrections.csv")

# Prep Dockets Data (Simple)

### read and review simple dockets dataset

df_dockets.head()

len(df_dockets)

df_dockets = df_dockets.rename(columns={'Docket Number':'docket','Violation Date':'violation_date',
                                        'Issuing Department Code':'dept',
                                        'Violation Address':'violation_address','Imposed Fine Detailed':'fine_amt'})
df_dockets.head()

### identify any dockets with missing violation date, which will be filtered out

df_dockets[df_dockets['violation_date'].isna()]

### filter for 7/1/19 to 6/30/23

#convert violation date to datetime, so it can be filtered
df_dockets['violation_date'] = pd.to_datetime(df_dockets['violation_date'])

df_dockets = df_dockets[(df_dockets['violation_date']>='7/1/19') & (df_dockets['violation_date']<='6/30/23')]

# get number of unique dockets, for reference
df_dockets['docket'].nunique()

# Prep Dockets Data (Detailed)

len(df_dockets_detailed)

# simplify field names
df_dockets_detailed = df_dockets_detailed.rename(columns={'Respondent Full Name':'respondent',
                                                          'Docket Number':'docket',
                                                         'Nov #':'violation_num',
                                                         'Violation Date':'violation_date',
                                                         'Hearing Date':'hearing_date',
                                                         'Violation Address':'violation_address',
                                                         'Violation Description':'violation_desc',
                                                         'Case Disposition Label':'case_disposition',
                                                         'Imposed Fine Detailed':'fine_amt'})
#df_dockets_detailed.head()

### filter detailed dockets for 7/1/19 to 6/30/23

#convert violation date to datetime, so it can be filtered
df_dockets_detailed['violation_date'] = pd.to_datetime(df_dockets_detailed['violation_date'])

df_dockets_detailed = df_dockets_detailed[(df_dockets_detailed['violation_date']>='7/1/19') 
                                          & (df_dockets_detailed['violation_date']<='6/30/23')]

len(df_dockets_detailed)

### identify data quality issues
Observation: the detailed docket dataset contains one docket (24DT000388) which is not assigned to any department

df_dockets_test_merge = pd.merge(df_dockets,df_dockets_detailed,on = "docket",how="right")

df_dockets_weird = df_dockets_test_merge[df_dockets_test_merge['dept'].isna()]
df_dockets_weird

### clean up fines data (clearing out $ helps with analysis)

# clean up fine data
df_dockets_detailed['fine_amt'] = pd.to_numeric(df_dockets_detailed['fine_amt'].replace('[\$,]', '', regex=True), errors='coerce')

# get number of unique dockets, for reference
df_dockets_detailed['docket'].nunique()

# Normalize Data

### dockets to departments
1917 records, because one record is missing department

df_dockets_to_depts = df_dockets[['docket','dept']].drop_duplicates()

df_dockets_to_depts.head()

len(df_dockets_to_depts)

### dockets to respondents
one docket can have multiple respondents. remove this for now as it could be calculated downstream during the data analysis phase

#df_dockets_to_respondents = df_dockets_detailed[['docket','respondent']].drop_duplicates()

#len(df_dockets_to_respondents)

### dockets to addresses

df_dockets_to_addresses = df_dockets_detailed[['docket','violation_address']].drop_duplicates()

df_dockets_to_addresses.head()

len(df_dockets_to_addresses)

# Prep Addresses for Geocoding

### clean up addresses

# add city and state, which are necessary for geocoding
df_dockets_to_addresses['city']='Chicago'
df_dockets_to_addresses['state']='Illinois'

# strip out trailing whitespace to get address merge to function. otherwise about 75% don't link correctly
df_dockets_to_addresses['violation_address'] = df_dockets_to_addresses['violation_address'].str.strip()

### merge with address corrections
I created the dataset "address-corrections" manually based on addresses which failed to geocode

df_corrections.head()

df_dockets_to_addresses_clean = pd.merge(df_dockets_to_addresses,df_corrections, on='violation_address', how='left')

# for correct records, copy existing address to cleaned address 
df_dockets_to_addresses_clean.loc[
    df_dockets_to_addresses_clean['corrected'].isna(), 'cleaned_address'] = df_dockets_to_addresses_clean['violation_address']
df_dockets_to_addresses_clean.head()

len(df_dockets_to_addresses_clean)

df_dockets_to_addresses_clean.groupby('corrected').size()

# sample some cleaned addresses
df_dockets_to_addresses_clean[df_dockets_to_addresses_clean['corrected']==1].head()

# Merge Department Info into Detailed Dockets Dataset
This gets all useful info in one place, albeit in denormalized form

len(df_dockets_detailed)

# left merge, to hold onto the record without a department assignment
df_dockets_detailed_dept = pd.merge(df_dockets_detailed,df_dockets_to_depts, on = 'docket', how = "left")
#df_dockets_detailed_dept.head()

len(df_dockets_detailed_dept)

df_dockets_detailed_dept['docket'].nunique()

# Export Data

df_dockets_to_depts.to_csv('../../data/01-tidied/dockets-to-depts.csv', index= False)
df_dockets_to_addresses_clean.to_csv('../../data/01-tidied/dockets-to-addresses.csv', index = False)
df_dockets_detailed_dept.to_csv('../../data/01-tidied/dockets-to-details.csv', index = False)
#df_dockets_to_respondents.to_csv('../../data/01-tidied/dockets-to-respondents.csv', index = False)
