==> notebooks/01-data-preparation/01d-tidy-dss-citations.ipynb <==
import pandas as pd

df=pd.read_csv("../../data/00-raw/FOIA-F067484-021424-snow-citations-streets-and-san.csv")

df.rename(columns={'INSPECTOR NAME':'inspector','VIOLATION CODE':'violation_code',
                   'DATE OF VIOLATION':'violation_date','DOCKET NUMBER':'docket',
                   'DISPOSITION':'disposition','ADDRESS':'citation_address','WARD':'citation_ward'}, inplace=True)
df = df[['inspector','violation_code','violation_date','docket','disposition','citation_address','citation_ward']]
df['violation_date']=pd.to_datetime(df['violation_date'])

len(df)

df1923 = df[(df['violation_date']>='7/1/19') & 
           (df['violation_date']<='6/30/23')]

len(df1923)

# Export

df1923.to_csv('../../data/01-tidied/dss-citations.csv', index=False)
