==> notebooks/01-data-preparation/02-geocode-addresses.ipynb <==
# Geocode Addresses
3/2/24
I completed geocoding on my laptop using QGIS and not Python.
<div style="color:red">this notebook should be cleaned up a bit</div>

### Preparation
Because geocoding takes time, I deduplicated the address list in data/01-tidied/dockets-to-addresses.csv so that each address only needs to be geocoded once.

### Step 1: Geocode Fines Data (in QGIS)

<ol>
<li>Drag and drop dockets-to-addresses.csv into layers (I added a city and state field here for Geocoding)
<li>MMQGIS -> Geocode > Geocode CSV with Web Service
<li>In the Web Service Geocode box:
    
<ul>
    <li>Input CSV file: dockets-to-addresses.csv
<li>Address: Cleaned Address
<li>City
<li>State
<li>Web Service: OpenStreetMap; this is free but may have usage limits.
<li>*Output File Name:*
<li>data\02-geocoded\dockets-geocoded.geojson
type = geoJSON
<li>Not Found Output List
<li>data\02-geocoded\dockets-not-geocoded.csv
<li>Apply
    </ul>
    
<li>Wait. This takes something like an hour for 2,000 records
<li>Geocoded layer (fines-geocoded in this case) should appear under Layers
Right-click -> attribute table
Should see a column latlong containing both latitude and longitude coordinates for each address
<Br>


### Step 2: Review Non-Geocoded Records
On the first take, geocoding was unable to process several records, due largely to incorrect addresses. I created a file 00-raw/address-corrections.csv to cross-walk bad addresses with cleaned addresses, and then re-ran 01c-tidy-unshoveled-sidewalk-fines.ipynb

| correction type |  example |
|---|---|
| missing street type | | added "AVE" to W BEACH, W GRAND, S CENTRAL, etc. |
| incorrect directional | 3430 W BELL = 3430 N BELL because S BELL doesn't exist |
| incorrect street type | SEELEY ST = SEELEY AVE |
| incomplete street name | MARTIN LUTH DR = MARTIN LUTHER KING JR DR |
| parsing problem | Des Plaines = Desplaines |
| typos/spelling | Lawerence = Lawrence |

4 unknown records<br>
84 address corrections

# Export

This results in two files:
* *fines-geocoded.geojson*- this file is used for spatially joining addresses
* *fines-not-geocoded.csv*- this file informs iterative address cleaning (see step 2)
