==> notebooks/01-data-preparation/03-spatially-join-addresses.ipynb <==
# Spatially Join Addresses
3/2/24
<div style="color:red">this notebook should be cleaned up a bit</div>
repeat with Communities, then 2015 Wards

# Input

data/01-tidied/dockets-to-addresses.csv

# Spatial Join

<ol>
<li>Drag and drop “Boundaries - Community Areas (current)” into layers
<li>Drag and drop dockets-geocoded.geojson into layers
<li>Vector > Data Management Tools > Join Attributes by Location.

<ul>
<li>Join to features in: dockets-geocoded (points layer)
<li>Features they: are within
<li>By comparing to: Boundaries - Community Areas
<li>Join type: Take attributes of the first matching feature only (one-to-one)
</ul>
    
<li>click Run
<li>click close
<li>right-click “Joined layer” -> Export -> Saved Features As…
format -> CSV
file name: 03-spatially-joined/addresses-to-communities.csv<br>
geometry-> no geometry
<li>OK
    </ol>

# Output

This results in two files:
* *addresses-to-communities.csv*- maps each address to a community area
* *addresses-to-wards2015.csv*- maps each address to a ward
