==> notebooks/01-data-preparation/04-standardize-data-with-gis.ipynb <==
# Standardize Snow Clearance Data
last updated 3/3/24

<a name="read"></a>
# 1. Read Spatially Joined Data

import pandas as pd

df_communities = pd.read_csv("../../data/03-spatially-joined/addresses-to-communities.csv")
df_wards = pd.read_csv("../../data/03-spatially-joined/addresses-to-wards2015.csv")
df_dockets = pd.read_csv("../../data/01-tidied/dockets-to-details.csv")

# four dockets with unknown address are omitted because they're not geocoded
len(df_communities)

# four dockets with unknown address are omitted because they're not geocoded
len(df_wards)

len(df_dockets)

df_dockets['docket'].nunique()

# Merge Geographies

df_communities = df_communities[['docket','cleaned_address','latlong','community']]
df_communities.head()

df_wards = df_wards[['docket','ward']]
df_wards.dropna(subset=['ward'], inplace=True)
df_wards['ward'] = df_wards['ward'].astype(int).astype(str)
df_wards = df_wards.rename(columns={'ward':'ward_1523'})
df_wards.head()

### merge address data

#left-hand join b/c we want to include records with no ward assignment
df_gis = pd.merge(df_communities,df_wards,on="docket", how = 'left')
df_gis.head()

len(df_gis)

### parse lat and long

#parse lat and long
df_gis['lat']=df_gis['latlong'].str.split(',').str[0]
df_gis['long']=df_gis['latlong'].str.split(',').str[1]

### review and clean records with no community assigned

df_gis[df_gis["community"].isna()]

df_gis[df_gis["ward_1523"].isna()]

One address, 5450 S 47TH (docket 21DT002747), is not a valid address but was geocoded at 41.8654712,-87.7423005 at the edge of Lawndale by Roosevelt, which doesn't look right. If it's 5450 W. 47th it'd be in suburban Stickney. So I'm choosing to skip it.

#drop record with invalid address
# index_to_drop = df[df['address'] == '5450 S 47TH'].index
# df = df.drop(index_to_drop)

# manually assign community areas for the other two addresses. strangely this address isn't recognized, and placed in West Garfield if I use BLVD instead of ST
df_gis.loc[df_gis['cleaned_address'] == '300 W WASHINGTON ST', 'community'] = 'LOOP'

<a name="docket"></a>
# Merge Geographies into Detailed Dockets

df_dockets_gis = pd.merge(df_dockets,df_gis,on='docket', how = 'left')

len(df_dockets_gis)

df_dockets_gis['docket'].nunique()

### identify and address missing values before grouping

df_dockets_gis[df_dockets_gis["community"].isna()]

df_dockets_gis[df_dockets_gis["ward_1523"].isna()]

df_dockets_gis[df_dockets_gis["dept"].isna()]

df_dockets_gis['dept']=df_dockets_gis['dept'].fillna('unknown')
df_dockets_gis['ward_1523']=df_dockets_gis['ward_1523'].fillna('unknown')
df_dockets_gis['community']=df_dockets_gis['community'].fillna('unknown')
df_dockets_gis['lat']=df_dockets_gis['lat'].fillna('unknown')
df_dockets_gis['long']=df_dockets_gis['long'].fillna('unknown')

# Docket-Level Summary
I expect 1918 dockets

# by docket
df_dockets_summary=df_dockets_gis.groupby(['docket','dept','violation_address','lat','long','community','ward_1523','violation_date']).agg(
    n_records=('docket','count'),
).reset_index()

df_dockets_summary.head()

len(df_dockets_summary)

# 5. Export

df_gis.to_csv("../../data/04-standardized/dockets-to-gis.csv", index=False)
df_dockets_gis.to_csv("../../data/04-standardized/dockets-to-details-gis.csv", index= False)
df_dockets_summary.to_csv("../../data/04-standardized/dockets-summary.csv", index= False)
