==> notebooks/01-data-preparation/05-move-finalized-datasets.ipynb <==
# Copy Finalized Datasets
3/3/24
This repo simply moves all finalized datasets into one place

import shutil

# snow complaints and citations from 01-tidied
shutil.copy('../../data/01-tidied/311-complaints-snow-all-types.csv',
            '../../data/05-finalized/311-complaints-snow-all-types.csv')
shutil.copy('../../data/01-tidied/311-complaints-unshoveled.csv',
            '../../data/05-finalized/311-complaints-unshoveled.csv')
shutil.copy('../../data/01-tidied/census-by-community.csv',
            '../../data/05-finalized/census-by-community.csv')
shutil.copy('../../data/01-tidied/dss-citations.csv',
            '../../data/05-finalized/dss-citations.csv')

# dockets data from 04-standardized
shutil.copy('../../data/04-standardized/dockets-to-details-gis.csv',
            '../../data/05-finalized/dockets-to-details-gis.csv')

shutil.copy('../../data/04-standardized/dockets-summary.csv',
            '../../data/05-finalized/dockets-summary.csv')

# geographic aggregates from 05-aggregated
# shutil.copy('../../data/05-aggregated/dockets-summary.csv',
#             '../../data/06-finalized/dockets-summary.csv')
# shutil.copy('../../data/05-aggregated/dockets-summary.csv',
#             '../../data/06-finalized/dockets-by-address.csv')
# shutil.copy('../../data/05-aggregated/dockets-summary.csv',
#             '../../data/06-finalized/dockets-by-community.csv')
# shutil.copy('../../data/05-aggregated/dockets-summary.csv',
#             '../../data/06-finalized/dockets-by-ward.csv')
